+++ docs/conf.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 import sys
 import os
 import datetime as dt
@@ -20,8 +19,8 @@ source_suffix = ".rst"
 master_doc = "index"
 
 # General information about the project.
-project = u"doitlive"
-copyright = u"2014-{0:%Y}".format(dt.datetime.utcnow())
+project = "doitlive"
+copyright = f"2014-{dt.datetime.utcnow():%Y}"
 
 version = release = doitlive.__version__