+++ tests/test_python_consoles.py
@@ -18,7 +18,7 @@ class TestPlayerConsole:
     )
     def test_interact(self, runner, console, command, expected):
         console.commands = [command]
-        with runner.isolation(input="{}\n\n".format(command)) as (stdout, stderr):
+        with runner.isolation(input=f"{command}\n\n") as (stdout, stderr):
             console.interact()
         assert expected in stdout.getvalue()